Moved config and database handling into sub-crates (see desc)
Reasoning is that they will be used in both the game server and the web server. In addition, I am thinking of a gateway to sit in front of both -- which would also need access to these.

While they are very small sub-crates, logically it makes sense to separate them out as they will talk with stand-alone binaries and libraries alike.